refactor(auth): rename auth_jwt → bot_provision + add status edge case
Phase 4 polish.

#1 module rename: octo-server/modules/auth_jwt → modules/bot_provision.
After Phase 4 stripped JWT signing/JWKS/exchange logic, the module only
holds mintBot + botToken (bot provisioning endpoints) — name was misleading.

- `git mv` modules/auth_jwt → modules/bot_provision
- All 5 files: package auth_jwt → package bot_provision
- Type AuthJWT → BotProvision (matches new module name)
- internal/modules.go import path updated
- log tag "AuthJWT" → "BotProvision"

#8 test coverage: verify-api-key handler test adds NonActiveStatus case
(space_member.status=2). Same SQL path as existing OwnerLeftSpace (status=0)
test — WHERE status=1 rejects both — but explicit case prevents regression
if someone later loosens to `status != 0`.

Build clean. Unit tests pending fresh test DB (current DB污染 from web
e2e runs, unrelated to this commit).
