+++ modules/bot_provision/1module.go
@@ -1,4 +1,4 @@
-package auth_jwt
+package bot_provision
 
 import (
 	"github.com/Mininglamp-OSS/octo-lib/config"
